move boundary condition conversion tests from `test_converters.py` to the `test_converter_boundary_condition.py`
