+++ tests/tools/test_converter_boundary_condition.py
@@ -1,8 +1,92 @@
+from pathlib import Path
 from typing import Dict
 from unittest.mock import Mock
 
+from hydrolib.core.basemodel import DiskOnlyFileModel
+from hydrolib.core.dflowfm.ext.models import Boundary
+from hydrolib.core.dflowfm.extold.models import ExtOldForcing, ExtOldQuantity
+from hydrolib.tools.extforce_convert.converters import BoundaryConditionConverter
 from hydrolib.tools.extforce_convert.main_converter import ExternalForcingConverter
 from tests.utils import compare_two_files
+
+
+class TestBoundaryConverter:
+
+    def test_merge_tim_files(self, input_files_dir: Path):
+        """
+        Test merging multiple tim files into a single tim model.
+        """
+        file_name = input_files_dir / "boundary-conditions/tfl_01.pli"
+        path_list = [
+            input_files_dir / "boundary-conditions/tfl_01_0001.tim",
+            input_files_dir / "boundary-conditions/tfl_01_0002.tim",
+        ]
+        forcing = ExtOldForcing(
+            quantity=ExtOldQuantity.WaterLevelBnd,
+            filename=file_name,
+            filetype=9,
+            method="3",
+            operand="O",
+        )
+        tim_model = BoundaryConditionConverter.merge_tim_files(path_list, forcing)
+        assert tim_model.quantities_names == ["tfl_01_0001", "tfl_01_0002"]
+        df = tim_model.as_dataframe()
+        assert df.index.tolist() == [0, 120]
+        assert df.columns.tolist() == ["tfl_01_0001", "tfl_01_0002"]
+        assert df.values.tolist() == [[0.01, 0.01], [0.01, 0.01]]
+
+    def test_with_pli(self, input_files_dir):
+        """
+        Old quantity block:
+
+        ```
+        QUANTITY =waterlevelbnd
+        FILENAME =tfl_01.pli
+        FILETYPE =9
+        METHOD   =3
+        OPERAND  =O
+        ```
+        """
+        file_name = input_files_dir / "boundary-conditions/tfl_01.pli"
+        forcing = ExtOldForcing(
+            quantity=ExtOldQuantity.WaterLevelBnd,
+            filename=file_name,
+            filetype=9,  # "Polyline"
+            method="3",  # "Interpolate space",
+            operand="O",
+        )
+
+        converter = BoundaryConditionConverter()
+        converter.root_dir = input_files_dir / "boundary-conditions"
+        start_date = "minutes since 2015-01-01 00:00:00"
+        new_quantity_block = converter.convert(forcing, start_date)
+        assert isinstance(new_quantity_block, Boundary)
+        assert new_quantity_block.quantity == "waterlevelbnd"
+        forcing_model = new_quantity_block.forcingfile
+        assert new_quantity_block.locationfile == DiskOnlyFileModel(file_name)
+        assert new_quantity_block.nodeid is None
+        assert new_quantity_block.bndwidth1d is None
+        assert new_quantity_block.bndbldepth is None
+        assert len(forcing_model.forcing) == 2
+        names = ["L1_0001", "L1_0002"]
+        assert all(
+            forcing.name == name for forcing, name in zip(forcing_model.forcing, names)
+        )
+        assert all(
+            [
+                forcing_model.forcing[i].quantityunitpair[0].quantity == "time"
+                for i in range(2)
+            ]
+        )
+        # check forcing model file path
+        assert forcing_model.filepath.name == "tfl_01.bc"
+        assert all(
+            [
+                forcing_model.forcing[i].quantityunitpair[1].quantity == "waterlevelbnd"
+                for i in range(2)
+            ]
+        )
+        assert forcing_model.forcing[0].datablock == [[0, 0.01], [120, 0.01]]
 
 
 def test_converter_update(old_forcing_file_boundary: Dict[str, str]):
